refactor(dao-all): 将 DeptTreeVo 中的 level 字段类型从 int 改为 long- 修改了 DeptTreeVo 类中的 level 字段类型，以适应可能的组织结构扩展 - 此修改可能会影响使用该字段的相关代码和数据库查询

+++ common-dao-module-all/dao-all/src/main/java/com/minimalism/vo/dept/DeptTreeVo.java
@@ -17,5 +17,5 @@ import java.io.Serializable;
 public class DeptTreeVo implements Serializable {
     private Long id;
     private Long parentId;
-    private int level;
+    private long level;
 }
